+++ app/Models/RequestErrorLog.php
@@ -4,10 +4,12 @@ declare(strict_types=1);
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Query\Builder as QueryBuilder;
 
 /**
  * Request error log
  *
+ * @mixin QueryBuilder
  * @author Dean Blackborough <dean@g3d-development.com>
  * @copyright Dean Blackborough 2018-2019
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
@@ -23,8 +25,23 @@ class RequestErrorLog extends Model
         return count($this->select('id')->get());
     }
 
-    public function paginatedCollection(int $offset = 0, int $limit = 10)
+    public function paginatedCollection(
+        int $offset = 0,
+        int $limit = 10
+    ): array
     {
-        return $this->orderByDesc('created_at')->offset($offset)->limit($limit)->get();
+        return $this->select(
+                'request_error_log.method AS request_error_log_method',
+                'request_error_log.expected_status_code AS request_error_log_expected_status_code',
+                'request_error_log.returned_status_code AS request_error_log_returned_status_code',
+                'request_error_log.request_uri AS request_error_log_request_uri',
+                'request_error_log.source AS request_error_log_source',
+                'request_error_log.created_at AS request_error_log_created_at'
+            )->
+            orderByDesc('created_at')->
+            offset($offset)->
+            limit($limit)->
+            get()->
+            toArray();
     }
 }